refactor data extractor
