+++ tests/data/test_utils.py
@@ -14,41 +14,6 @@ from lightning_pose.data.utils import (
 )
 
 
-def test_data_extractor(base_data_module_combined, multiview_heatmap_data_module_combined):
-
-    from lightning_pose.data.utils import DataExtractor
-
-    # ---------------------------
-    # supervised single view
-    # ---------------------------
-    num_frames = (
-        len(base_data_module_combined.dataset)
-        * base_data_module_combined.train_probability
-    )
-    keypoint_tensor, _ = DataExtractor(
-        data_module=base_data_module_combined, cond="train"
-    )()
-    assert keypoint_tensor.shape == (num_frames, 34)  # 72 = 0.8 * 90 images, 17 * 2 coordinates
-
-    keypoint_tensor, images_tensor = DataExtractor(
-        data_module=base_data_module_combined, cond="train", extract_images=True
-    )()
-    assert images_tensor is not None
-    assert images_tensor.shape == (num_frames, 3, 128, 128)
-
-    # ---------------------------
-    # supervised multiview
-    # ---------------------------
-    num_frames = (
-        len(multiview_heatmap_data_module_combined.dataset)
-        * multiview_heatmap_data_module_combined.train_probability
-    )
-    keypoint_tensor, _ = DataExtractor(
-        data_module=multiview_heatmap_data_module_combined, cond="train"
-    )()
-    assert keypoint_tensor.shape == (num_frames, 28)  # 72 = 0.8 * 90 images, 7 * 2 * 2 coords
-
-
 def test_split_sizes_from_probabilities():
 
     from lightning_pose.data.utils import split_sizes_from_probabilities
